Modify timestamps based on git commit time
I asked this question on

Based on https://github.com/quarto-dev/quarto-cli/discussions/5899

The suggestion was to use `freeze: auto`, but it wasn't clear to me how
that would work. First, I would need to commit `_freeze` directory to
git but that is tricky because I don't run quarto locally. I was also
not sure what will hapen to cross-links if I do so (think about chapter
number changing).

Then I found a script to change modification times.

https://stackoverflow.com/questions/21735435/git-clone-changes-file-modification-time/55609950#55609950

After a bit of trial and error it is now working correctly.